+++ packages/task/src/browser/task-frontend-module.ts
@@ -29,10 +29,14 @@ import { TaskServer, taskPath } from '../common/task-protocol';
 import { TaskWatcher } from '../common/task-watcher';
 import { bindProcessTaskModule } from './process/process-task-frontend-module';
 import { TaskSchemaUpdater } from './task-schema-updater';
+import { TaskActionProvider, ConfigureTaskAction } from './task-action-provider';
+import '../../src/browser/style/index.css';
 
 export default new ContainerModule(bind => {
     bind(TaskFrontendContribution).toSelf().inSingletonScope();
     bind(TaskService).toSelf().inSingletonScope();
+    bind(TaskActionProvider).toSelf().inSingletonScope();
+    bind(ConfigureTaskAction).toSelf().inSingletonScope();
 
     for (const identifier of [FrontendApplicationContribution, CommandContribution, KeybindingContribution, MenuContribution, QuickOpenContribution]) {
         bind(identifier).toService(TaskFrontendContribution);